# openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "My API",
      "version": "1.0.0",
      "description": "Comprehensive API documentation"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
      "/api/v1/branches": {
        "post": {
          "summary": "Create a new branch",
          "tags": [
            "Branches"
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "1234567890"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Branch created successfully"
            }
          }
        },
        "get": {
          "summary": "Get all branches",
          "tags": [
            "Branches"
          ],
          "responses": {
            "200": {
              "description": "List of all branches"
            }
          }
        }
      },
      "/api/v1/branches/{id}": {
        "get": {
          "summary": "Get a branch by ID",
          "tags": [
            "Branches"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the branch",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Branch details"
            }
          }
        },
        "put": {
          "summary": "Update a branch by ID",
          "tags": [
            "Branches"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the branch",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "1234567890"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Branch updated successfully"
            }
          }
        },
        "delete": {
          "summary": "Delete a branch by ID",
          "tags": [
            "Branches"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the branch",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Branch deleted successfully"
            }
          }
        }
      },
      "/api/v1/employees": {
        "post": {
          "summary": "Create a new employee",
          "tags": [
            "Employees"
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Employee created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            },
            "400": {
              "description": "Bad request due to validation error"
            }
          }
        },
        "get": {
          "summary": "Get all employees",
          "tags": [
            "Employees"
          ],
          "responses": {
            "200": {
              "description": "List of employees",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Employee"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/employees/{id}": {
        "get": {
          "summary": "Get an employee by ID",
          "tags": [
            "Employees"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Employee details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            },
            "404": {
              "description": "Employee not found"
            }
          }
        },
        "put": {
          "summary": "Update an employee by ID",
          "tags": [
            "Employees"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Employee updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            },
            "400": {
              "description": "Validation error"
            },
            "404": {
              "description": "Employee not found"
            }
          }
        },
        "delete": {
          "summary": "Delete an employee by ID",
          "tags": [
            "Employees"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Employee deleted successfully"
            },
            "404": {
              "description": "Employee not found"
            }
          }
        }
      },
      "/api/v1/employees/branch/{branchId}": {
        "get": {
          "summary": "Get employees by branch ID",
          "tags": [
            "Employees"
          ],
          "parameters": [
            {
              "name": "branchId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of employees in the specified branch",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Employee"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/employees/department/{department}": {
        "get": {
          "summary": "Get employees by department",
          "tags": [
            "Employees"
          ],
          "parameters": [
            {
              "name": "department",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of employees in the specified department",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Employee"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/health": {
        "get": {
          "summary": "Check the health of the server",
          "description": "Returns a simple message indicating the server is healthy",
          "responses": {
            "200": {
              "description": "Server is healthy",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Server is healthy"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Employee": {
          "type": "object",
          "required": [
            "name",
            "email"
          ],
          "properties": {
            "id": {
              "type": "string",
              "example": "emp123"
            },
            "name": {
              "type": "string",
              "example": "Shubh Rapper"
            },
            "position": {
              "type": "string",
              "example": "Software Engineer"
            },
            "department": {
              "type": "string",
              "example": "Engineering"
            },
            "email": {
              "type": "string",
              "example": "shubh.rapper@example.com"
            },
            "phone": {
              "type": "string",
              "example": "1234567890"
            },
            "branchId": {
              "type": "string",
              "example": "branch123"
            }
          }
        }
      }
    },
    "tags": []
  }